+++ hasOwnProperty.ts
@@ -1,3 +1,10 @@
-export default function hasOwnProperty<T>(obj: T, propName: string | number | symbol): propName is keyof T {
-  return Object.prototype.hasOwnProperty.call(obj, propName);
+const objectHasOwnProperty = Object.prototype.hasOwnProperty;
+
+function hasOwnProperty(obj: null | undefined, propName: PropertyKey): false;
+function hasOwnProperty<T, K extends PropertyKey>(obj: T, propName: K): obj is T & Record<K, unknown>;
+function hasOwnProperty<T, K extends PropertyKey>(obj: T, propName: K): propName is K & keyof T;
+function hasOwnProperty<T, K extends PropertyKey>(obj: T, propName: K): propName is K & keyof T {
+  return obj != null && objectHasOwnProperty.call(obj, propName);
 }
+
+export default hasOwnProperty;